requirements.txt
olefile
skidl
schemdraw
pillow
